+++ 3-java-extension.sh
@@ -4,7 +4,7 @@ cd "$ROOT"
 source "$ROOT/demo-magic/demo-magic.sh"
 
 DEMO_PROMPT="${GREEN}➜ ${CYAN}\W ${COLOR_RESET}"
-TYPE_SPEED=30
+TYPE_SPEED=50
 
 function comment() {
   cmd=$DEMO_COMMENT_COLOR$1$COLOR_RESET
@@ -30,7 +30,7 @@ pe
 osascript -e "tell application \"Terminal\" to do script \"cd $ROOT/opensearch-sdk-java; ./gradlew helloWorld\""
 
 comment "Here's the OpenSearch we're running:"
-pe "curl http://localhost:9200"
+pe "curl http://localhost:9200 | jq"
 
 comment "Install the Java Extension into OpenSearch:"
 pe "curl -s -XPOST \"localhost:9200/_extensions/initialize\" -H \"Content-Type:application/json\" --data @src/main/java/org/opensearch/sdk/sample/helloworld/hello.json | jq"
